Break down nucleus reducers into separate files
