+++ core/src/nucleus/mod.rs
@@ -1,23 +1,18 @@
 /// Nucleus is the module that handles DNA, including the Ribosome.
 ///
 pub mod actions;
+pub mod reducers;
 pub mod ribosome;
 pub mod state;
 
 use crate::{
-    action::{Action, ActionWrapper, NucleusReduceFn},
-    context::Context,
+    action::{Action, ActionWrapper},
     instance::Observer,
-    nucleus::{
-        ribosome::api::call::{reduce_call, validate_call},
-        state::{NucleusState, NucleusStatus},
-    },
 };
 use holochain_core_types::{
     cas::content::Address,
     dna::{
         capabilities::{CapabilityCall, CapabilityType},
-        wasm::DnaWasm,
         Dna,
     },
     error::{HcResult, HolochainError},
@@ -25,13 +20,12 @@ use holochain_core_types::{
 };
 use snowflake;
 use std::{
-    sync::{
-        mpsc::{channel, SyncSender},
-        Arc,
-    },
-    thread,
+    sync::mpsc::channel,
     time::Duration,
 };
+
+pub use crate::nucleus::reducers::execute_zome_function::ExecuteZomeFnResponse;
+pub use crate::nucleus::reducers::reduce;
 
 /// Struct holding data for requesting the execution of a Zome function (ExecutionZomeFunction Action)
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
@@ -75,23 +69,7 @@ impl ZomeFnCall {
     }
 }
 
-/// WIP - Struct for holding data when requesting an Entry Validation (ValidateEntry Action)
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct EntrySubmission {
-    pub zome_name: String,
-    pub type_name: String,
-    pub entry_content: String,
-}
-
-impl EntrySubmission {
-    pub fn new<S: Into<String>>(zome_name: S, type_name: S, content: S) -> Self {
-        EntrySubmission {
-            zome_name: zome_name.into(),
-            type_name: type_name.into(),
-            entry_content: content.into(),
-        }
-    }
-}
+pub type ZomeFnResult = HcResult<JsonString>;
 
 /// Dispatch ExecuteZoneFunction to Instance and block until call has finished.
 /// for test only?? <-- (apparently not, since it's used in Holochain::call)
@@ -115,235 +93,6 @@ pub fn call_and_wait_for_result(
         } else {
             let _ = tick_rx.recv_timeout(Duration::from_millis(100));
         }
-    }
-}
-
-pub type ZomeFnResult = HcResult<JsonString>;
-
-#[derive(Clone, Debug, PartialEq, Hash)]
-pub struct ExecuteZomeFnResponse {
-    call: ZomeFnCall,
-    result: ZomeFnResult,
-}
-
-impl ExecuteZomeFnResponse {
-    pub fn new(call: ZomeFnCall, result: Result<JsonString, HolochainError>) -> Self {
-        ExecuteZomeFnResponse { call, result }
-    }
-
-    /// read only access to call
-    pub fn call(&self) -> ZomeFnCall {
-        self.call.clone()
-    }
-
-    /// read only access to result
-    pub fn result(&self) -> Result<JsonString, HolochainError> {
-        self.result.clone()
-    }
-}
-
-/// Reduce ReturnInitializationResult Action
-/// On initialization success, set Initialized status
-/// otherwise set the failed message
-#[allow(unknown_lints)]
-#[allow(needless_pass_by_value)]
-fn reduce_return_initialization_result(
-    _context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    if state.status() != NucleusStatus::Initializing {
-        state.status = NucleusStatus::InitializationFailed(
-            "reduce of ReturnInitializationResult attempted when status != Initializing".into(),
-        );
-    } else {
-        let action = action_wrapper.action();
-        let result = unwrap_to!(action => Action::ReturnInitializationResult);
-        match result {
-            None => state.status = NucleusStatus::Initialized,
-            Some(err) => state.status = NucleusStatus::InitializationFailed(err.clone()),
-        };
-    }
-}
-
-/// Reduce InitApplication Action
-/// Switch status to failed if an initialization is tried for an
-/// already initialized, or initializing instance.
-#[allow(unknown_lints)]
-#[allow(needless_pass_by_value)]
-fn reduce_init_application(
-    _context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    match state.status() {
-        NucleusStatus::Initializing => {
-            state.status =
-                NucleusStatus::InitializationFailed("Nucleus already initializing".to_string())
-        }
-        NucleusStatus::Initialized => {
-            state.status =
-                NucleusStatus::InitializationFailed("Nucleus already initialized".to_string())
-        }
-        NucleusStatus::New | NucleusStatus::InitializationFailed(_) => {
-            let ia_action = action_wrapper.action();
-            let dna = unwrap_to!(ia_action => Action::InitApplication);
-            // Update status
-            state.status = NucleusStatus::Initializing;
-            // Set DNA
-            state.dna = Some(dna.clone());
-        }
-    }
-}
-
-pub(crate) fn launch_zome_fn_call(
-    context: Arc<Context>,
-    zome_call: ZomeFnCall,
-    wasm: &DnaWasm,
-    dna_name: String,
-) {
-    let code = wasm.code.clone();
-
-    thread::spawn(move || {
-        // Have Ribosome spin up DNA and call the zome function
-        let call_result = ribosome::run_dna(
-            &dna_name,
-            context.clone(),
-            code,
-            &zome_call,
-            Some(zome_call.clone().parameters.into_bytes()),
-        );
-        // Construct response
-        let response = ExecuteZomeFnResponse::new(zome_call.clone(), call_result);
-        // Send ReturnZomeFunctionResult Action
-        context
-            .action_channel()
-            .send(ActionWrapper::new(Action::ReturnZomeFunctionResult(
-                response,
-            )))
-            .expect("action channel to be open in reducer");
-    });
-}
-
-/// Reduce ExecuteZomeFunction Action
-/// Execute an exposed Zome function in a separate thread and send the result in
-/// a ReturnZomeFunctionResult Action on success or failure
-fn reduce_execute_zome_function(
-    context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    let fn_call = match action_wrapper.action().clone() {
-        Action::ExecuteZomeFunction(call) => call,
-        _ => unreachable!(),
-    };
-
-    fn dispatch_error_result(
-        action_channel: &SyncSender<ActionWrapper>,
-        fn_call: &ZomeFnCall,
-        error: HolochainError,
-    ) {
-        let zome_not_found_response =
-            ExecuteZomeFnResponse::new(fn_call.clone(), Err(error.clone()));
-
-        action_channel
-            .send(ActionWrapper::new(Action::ReturnZomeFunctionResult(
-                zome_not_found_response,
-            )))
-            .expect("action channel to be open in reducer");
-    }
-
-    // 1. Validate the call (a number of things could go wrong)
-    let dna = match validate_call(context.clone(), state, &fn_call) {
-        Err(err) => {
-            // Notify failure
-            dispatch_error_result(context.action_channel(), &fn_call, err);
-            return;
-        }
-        Ok(dna) => dna,
-    };
-
-    // 2. function WASM and execute it in a separate thread
-    let maybe_code = dna.get_wasm_from_zome_name(fn_call.zome_name.clone());
-    let code =
-        maybe_code.expect("zome not found, Should have failed before when getting capability.");
-
-    // Ok Zome function is defined in given capability.
-    // Prepare call - FIXME is this really useful?
-    state.zome_calls.insert(fn_call.clone(), None);
-    // Launch thread with function call
-    launch_zome_fn_call(context, fn_call, &code, state.dna.clone().unwrap().name);
-}
-
-fn reduce_return_validation_result(
-    _context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    let action = action_wrapper.action();
-    let ((id, hash), validation_result) = unwrap_to!(action => Action::ReturnValidationResult);
-    state
-        .validation_results
-        .insert((id.clone(), hash.clone()), validation_result.clone());
-}
-
-/// Reduce ReturnZomeFunctionResult Action.
-/// Simply drops function call into zome_calls state.
-#[allow(unknown_lints)]
-#[allow(needless_pass_by_value)]
-fn reduce_return_zome_function_result(
-    _context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    let action = action_wrapper.action();
-    let fr = unwrap_to!(action => Action::ReturnZomeFunctionResult);
-    // @TODO store the action and result directly
-    // @see https://github.com/holochain/holochain-rust/issues/198
-    state.zome_calls.insert(fr.call(), Some(fr.result()));
-}
-
-fn reduce_return_validation_package(
-    _context: Arc<Context>,
-    state: &mut NucleusState,
-    action_wrapper: &ActionWrapper,
-) {
-    let action = action_wrapper.action();
-    let (id, maybe_validation_package) = unwrap_to!(action => Action::ReturnValidationPackage);
-    state
-        .validation_packages
-        .insert(id.clone(), maybe_validation_package.clone());
-}
-
-/// Maps incoming action to the correct reducer
-fn resolve_reducer(action_wrapper: &ActionWrapper) -> Option<NucleusReduceFn> {
-    match action_wrapper.action() {
-        Action::ReturnInitializationResult(_) => Some(reduce_return_initialization_result),
-        Action::InitApplication(_) => Some(reduce_init_application),
-        Action::ExecuteZomeFunction(_) => Some(reduce_execute_zome_function),
-        Action::ReturnZomeFunctionResult(_) => Some(reduce_return_zome_function_result),
-        Action::Call(_) => Some(reduce_call),
-        Action::ReturnValidationResult(_) => Some(reduce_return_validation_result),
-        Action::ReturnValidationPackage(_) => Some(reduce_return_validation_package),
-        _ => None,
-    }
-}
-
-/// Reduce state of Nucleus according to action.
-/// Note: Can't block when dispatching action here because we are inside the reduce's mutex
-pub fn reduce(
-    context: Arc<Context>,
-    old_state: Arc<NucleusState>,
-    action_wrapper: &ActionWrapper,
-) -> Arc<NucleusState> {
-    let handler = resolve_reducer(action_wrapper);
-    match handler {
-        Some(f) => {
-            let mut new_state: NucleusState = (*old_state).clone();
-            f(context, &mut new_state, &action_wrapper);
-            Arc::new(new_state)
-        }
-        None => old_state,
     }
 }
 
@@ -372,15 +121,13 @@ pub mod tests {
     extern crate test_utils;
     use super::*;
     use crate::{
-        action::{tests::test_action_wrapper_rzfr, ActionWrapper},
         instance::{
-            tests::{test_context, test_context_with_channels, test_instance},
+            tests::{test_context, test_instance},
             Instance,
         },
-        nucleus::state::tests::test_nucleus_state,
+        nucleus::state::{NucleusState, NucleusStatus},
     };
-    use holochain_core_types::dna::{capabilities::CapabilityCall, Dna};
-    use std::sync::{mpsc::sync_channel, Arc};
+    use holochain_core_types::dna::capabilities::CapabilityCall;
 
     use holochain_core_types::{
         error::DnaError,
@@ -473,95 +220,6 @@ pub mod tests {
     }
 
     #[test]
-    /// test for returning zome function result actions
-    fn test_reduce_return_zome_function_result() {
-        let context = test_context("jimmy", None);
-        let mut state = test_nucleus_state();
-        let action_wrapper = test_action_wrapper_rzfr();
-
-        // @TODO don't juggle action wrappers to get at action in state
-        // @see https://github.com/holochain/holochain-rust/issues/198
-        let action = action_wrapper.action();
-        let fr = unwrap_to!(action => Action::ReturnZomeFunctionResult);
-
-        reduce_return_zome_function_result(context, &mut state, &action_wrapper);
-
-        assert!(state.zome_calls.contains_key(&fr.call()));
-    }
-
-    #[test]
-    /// smoke test the init of a nucleus reduction
-    fn can_reduce_initialize_action() {
-        let dna = Dna::new();
-        let action_wrapper = ActionWrapper::new(Action::InitApplication(dna.clone()));
-        let nucleus = Arc::new(NucleusState::new()); // initialize to bogus value
-        let (sender, _receiver) = sync_channel::<ActionWrapper>(10);
-        let (tx_observer, _observer) = sync_channel::<Observer>(10);
-        let context = test_context_with_channels("jimmy", &sender, &tx_observer, None);
-
-        // Reduce Init action
-        let reduced_nucleus = reduce(context.clone(), nucleus.clone(), &action_wrapper);
-
-        assert_eq!(reduced_nucleus.has_initialized(), false);
-        assert_eq!(reduced_nucleus.has_initialization_failed(), false);
-        assert_eq!(reduced_nucleus.status(), NucleusStatus::Initializing);
-        assert!(reduced_nucleus.dna().is_some());
-        assert_eq!(reduced_nucleus.dna().unwrap(), dna);
-    }
-
-    #[test]
-    /// test that we can initialize and send/receive result values from a nucleus
-    fn can_reduce_return_init_result_action() {
-        let dna = Dna::new();
-        let action_wrapper = ActionWrapper::new(Action::InitApplication(dna));
-        let nucleus = Arc::new(NucleusState::new()); // initialize to bogus value
-        let (sender, _receiver) = sync_channel::<ActionWrapper>(10);
-        let (tx_observer, _observer) = sync_channel::<Observer>(10);
-        let context = test_context_with_channels("jimmy", &sender, &tx_observer, None).clone();
-
-        // Reduce Init action
-        let initializing_nucleus = reduce(context.clone(), nucleus.clone(), &action_wrapper);
-
-        assert_eq!(initializing_nucleus.has_initialized(), false);
-        assert_eq!(initializing_nucleus.has_initialization_failed(), false);
-        assert_eq!(initializing_nucleus.status(), NucleusStatus::Initializing);
-
-        // Send ReturnInit(false) ActionWrapper
-        let return_action_wrapper = ActionWrapper::new(Action::ReturnInitializationResult(Some(
-            "init failed".to_string(),
-        )));
-        let reduced_nucleus = reduce(
-            context.clone(),
-            initializing_nucleus.clone(),
-            &return_action_wrapper,
-        );
-
-        assert_eq!(reduced_nucleus.has_initialized(), false);
-        assert_eq!(reduced_nucleus.has_initialization_failed(), true);
-        assert_eq!(
-            reduced_nucleus.status(),
-            NucleusStatus::InitializationFailed("init failed".to_string())
-        );
-
-        // Reduce Init action
-        let reduced_nucleus = reduce(context.clone(), reduced_nucleus.clone(), &action_wrapper);
-
-        assert_eq!(reduced_nucleus.status(), NucleusStatus::Initializing);
-
-        // Send ReturnInit(None) ActionWrapper
-        let return_action_wrapper = ActionWrapper::new(Action::ReturnInitializationResult(None));
-        let reduced_nucleus = reduce(
-            context.clone(),
-            initializing_nucleus.clone(),
-            &return_action_wrapper,
-        );
-
-        assert_eq!(reduced_nucleus.has_initialized(), true);
-        assert_eq!(reduced_nucleus.has_initialization_failed(), false);
-        assert_eq!(reduced_nucleus.status(), NucleusStatus::Initialized);
-    }
-
-    #[test]
     /// tests that calling a valid zome function returns a valid result
     fn call_zome_function() {
         let dna = test_utils::create_test_dna_with_wat("test_zome", "test_cap", None);
@@ -579,21 +237,6 @@ pub mod tests {
 
         assert!(result.is_ok());
         assert_eq!(JsonString::from(RawString::from(1337)), result.unwrap());
-    }
-
-    #[test]
-    /// smoke test reducing over a nucleus
-    fn can_reduce_execfn_action() {
-        let call = ZomeFnCall::new("myZome", Some(test_capability_call()), "bogusfn", "");
-
-        let action_wrapper = ActionWrapper::new(Action::ExecuteZomeFunction(call));
-        let nucleus = Arc::new(NucleusState::new()); // initialize to bogus value
-        let (sender, _receiver) = sync_channel::<ActionWrapper>(10);
-        let (tx_observer, _observer) = sync_channel::<Observer>(10);
-        let context = test_context_with_channels("jimmy", &sender, &tx_observer, None);
-
-        let reduced_nucleus = reduce(context, nucleus.clone(), &action_wrapper);
-        assert_eq!(nucleus, reduced_nucleus);
     }
 
     #[test]
